ai-shopping-assistant/requirements.txt
fastapi
uvicorn[standard]
gunicorn
sqlalchemy
psycopg2-binary
apscheduler
requests
beautifulsoup4
python-dotenv
httpx
